pyproject.toml
[tool.poetry]
name = "llmagent"
version = "0.1.0"
description = ""
authors = ["dnarganes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.41.2"
tokenizers = "^0.19.1"
ipython = "^8.25.0"
pandas = "^2.2.2"
anthropic = "^0.28.1"
neo4j = "^5.21.0"
botocore = "^1.34.129"
boto3 = "^1.34.129"
streamlit = "^1.35.0"
typing-extensions = "^4.12.2"
openai = "0.28"
python-dotenv = "^1.0.1"
plotly = "^5.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = "py310"
include = '/src/.*\\.pyi?$|/tests/.*\\.pyi?$|/docs/.*\\.pyi?$|noxfile.py'

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
cache_dir = ".cache/pytest_cache"
minversion = "6.0"
addopts = "-rv --doctest-modules --color=yes"
testpaths = ['tests']
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(levelname)s:%(name)s: %(message)s"